pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "krnx-bench"
version = "0.1.0"
description = "KRNX Memory Kernel Benchmark Harness"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "KRNX Team"}
]
keywords = ["ai", "memory", "benchmark", "llm", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "docker>=7.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "qdrant-client>=1.7.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
krnx-bench = "cli:app"

[tool.hatch.build.targets.wheel]
packages = ["bench"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
